requirements.txt
flask
lxml
google-generativeai